Removed unused libraries package.
Tests have been moved in the corresponding internal package.
