+++ internal/libraries/repolist_test.go
@@ -59,3 +59,107 @@ func TestRepoFolderPathDetermination(t *testing.T) {
 	require.NoError(t, err)
 	require.Equal(t, "bitbucket.org/bjoern/arduino_osc", f)
 }
+
+func TestLoadRepoListFromFile(t *testing.T) {
+	_, err := LoadRepoListFromFile("./testdata/nonexistent.txt")
+	assert.Error(t, err, "Attempt to load non-existent registry data file")
+
+	repos, err := LoadRepoListFromFile("./testdata/git_test_repos.txt")
+	require.NoError(t, err)
+
+	reposAssertion := []*Repo{
+		{
+			URL:         "https://github.com/arduino-libraries",
+			Types:       []string{"Arduino"},
+			LibraryName: "libraries",
+		},
+		{
+			URL:         "git@github.com:PaulStoffregen/Audio.git",
+			Types:       []string{"Contributed"},
+			LibraryName: "Audio",
+		},
+		{
+			URL:         "https://github.com/PaulStoffregen/OctoWS2811.git",
+			Types:       []string{"Arduino", "Contributed"},
+			LibraryName: "OctoWS2811",
+		},
+		{
+			URL:         "https://github.com/PaulStoffregen/AltSoftSerial.git",
+			Types:       []string{"Contributed"},
+			LibraryName: "AltSoftSerial",
+		},
+		{
+			URL:         "https://github.com/Cheong2K/ble-sdk-arduino.git",
+			Types:       []string{"Contributed"},
+			LibraryName: "ble-sdk-arduino",
+		},
+		{
+			URL:         "https://github.com/arduino-libraries/Bridge.git",
+			Types:       []string{"Contributed"},
+			LibraryName: "Bridge",
+		},
+		{
+			URL:         "https://github.com/adafruit/Adafruit_ADS1X15.git",
+			Types:       []string{"Recommended"},
+			LibraryName: "Adafruit_ADS1X15",
+		},
+		{
+			URL:         "https://github.com/adafruit/Adafruit_ADXL345.git",
+			Types:       []string{"Recommended"},
+			LibraryName: "Adafruit_ADXL345",
+		},
+		{
+			URL:         "https://github.com/adafruit/Adafruit_AHRS.git",
+			Types:       []string{"Recommended"},
+			LibraryName: "Adafruit_AHRS",
+		},
+		{
+			URL:         "https://github.com/adafruit/Adafruit_AM2315.git",
+			Types:       []string{"Recommended"},
+			LibraryName: "Adafruit_AM2315",
+		},
+		{
+			URL:         "https://github.com/arduino-libraries/Scheduler.git",
+			Types:       []string{"Arduino"},
+			LibraryName: "Scheduler",
+		},
+		{
+			URL:         "https://github.com/arduino-libraries/SD.git",
+			Types:       []string{"Arduino"},
+			LibraryName: "SD",
+		},
+		{
+			URL:         "https://github.com/arduino-libraries/Servo.git",
+			Types:       []string{"Arduino"},
+			LibraryName: "Servo",
+		},
+	}
+
+	assert.Equal(t, reposAssertion, repos)
+}
+
+func TestListRepos(t *testing.T) {
+	repos, err := ListRepos("testdata/git_test_repos.txt")
+	require.Error(t, err)
+
+	require.Equal(t, 11, len(repos))
+
+	require.Equal(t, "https://github.com/PaulStoffregen/OctoWS2811.git", repos[0].URL)
+	require.Equal(t, "https://github.com/PaulStoffregen/AltSoftSerial.git", repos[1].URL)
+
+	require.Equal(t, "https://github.com/Cheong2K/ble-sdk-arduino.git", repos[2].URL)
+	require.Equal(t, "https://github.com/arduino-libraries/Bridge.git", repos[3].URL)
+	require.Equal(t, "https://github.com/adafruit/Adafruit_ADS1X15.git", repos[4].URL)
+	require.Equal(t, "https://github.com/adafruit/Adafruit_ADXL345.git", repos[5].URL)
+	require.Equal(t, "https://github.com/adafruit/Adafruit_AHRS.git", repos[6].URL)
+	require.Equal(t, "https://github.com/adafruit/Adafruit_AM2315.git", repos[7].URL)
+	require.Equal(t, "https://github.com/arduino-libraries/Scheduler.git", repos[8].URL)
+	require.Equal(t, "https://github.com/arduino-libraries/SD.git", repos[9].URL)
+	require.Equal(t, "https://github.com/arduino-libraries/Servo.git", repos[10].URL)
+	require.Error(t, err)
+
+	error, ok := err.(GitURLsError)
+	require.True(t, ok)
+	require.Equal(t, "https://github.com/arduino-libraries", error.Repos[0].URL)
+	require.Equal(t, "git@github.com:PaulStoffregen/Audio.git", error.Repos[1].URL)
+}
